+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
+### 7.3.0
+* Update tagging with category listing support
+* Add category export
 ### 7.2.6
 * Convert price in cart tagging item to avoid rounding errors